Added MAKE building support.
New issue
Summary: Add preferences support for $GOROOT, $GOOS, and $GOARCH.  Unless environment options are already provided somewhere.
If the environment does not specify GOROOT, GOOS, and/or GOARCH, they are set to the hard-coded values that work on my system.
This is probably not the behavior that we want.

git-svn-id: https://goclipse.googlecode.com/svn/trunk@6 1731708e-d1f7-11de-a1df-87ee4daaac7b